Use a struct instead of an u32 as instruction id
Capstone decodes the instruction's mnemonic into an integer (cs_insn.id). It
corresponds to a value of an arch-specific enum ([arch]_insn).

In C you can retrive the enum value by casting the instr's id.
For example if the current arch is arm:

  arm_insn mnemonic = (arm_insn)instr.id;

Unfortunately, in Rust it is not possible to cast integers into enums
without an unsafe block (std::mem::transmute).

Use an enum to represent the id. Each value is of type:
  ARCH_NAM(arch_specific_insn)

In this way, the value can obtained using a `if let` or a `match`.